chore: port RpcServerGuard
