+++ crates/rpc/src/lib.rs
@@ -12,7 +12,7 @@
 //! # use signet_rpc::{Pnt, RpcCtx};
 //! # use reth_node_api::FullNodeComponents;
 //! # use reth::tasks::TaskExecutor;
-//! use signet_rpc::{router, serve_axum};
+//! use signet_rpc::{router, ServeConfig};
 //!
 //! # pub async fn f<Host, Signet>(ctx: RpcCtx<Host, Signet>, tasks: &TaskExecutor) -> eyre::Result<()>
 //! # where
@@ -20,13 +20,17 @@
 //! #   Signet: Pnt,
 //! # {
 //! let router = signet_rpc::router().with_state(ctx);
-//! // Spawn the server on the given addresses.
-//! let _ = serve_axum(
-//!     tasks,
-//!     &router,
-//!     &["localhost:8080".parse()?],
-//!     None
-//! ).await?;
+//!
+//! let cfg = ServeConfig {
+//!     http: vec!["localhost:8080".parse()?],
+//!     ws: vec![],
+//!     cors: None,
+//!     ipc: None,
+//! };
+//!
+//! // Spawn the server on the given addresses, the shutdown guard
+//! // will shutdown the server(s) when dropped.
+//! let shutdown_guard = cfg.serve(tasks, router).await?;
 //! # Ok(())
 //! # }
 //! ```
@@ -43,6 +47,9 @@
 #![deny(unused_must_use, rust_2018_idioms)]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
+mod config;
+pub use config::{RpcServerGuard, ServeConfig};
+
 mod ctx;
 pub use ctx::RpcCtx;
 
@@ -57,24 +64,15 @@ mod interest;
 mod forwarder;
 use forwarder::TxCacheForwarder;
 
-pub(crate) mod util;
+/// Utils and simple serve functions.
+pub mod util;
 pub use util::Pnt;
 
 /// Re-exported for convenience
 pub use ::ajj;
 
-use ajj::{
-    pubsub::{Connect, ServerShutdown},
-    Router,
-};
-use axum::http::{HeaderValue, Method};
-use interprocess::local_socket as ls;
-use reth::tasks::TaskExecutor;
+use ajj::Router;
 use reth_node_api::FullNodeComponents;
-use std::{future::IntoFuture, net::SocketAddr};
-use tokio::task::JoinHandle;
-use tower_http::cors::{AllowOrigin, CorsLayer};
-use tracing::error;
 
 /// Create a new router with the given host and signet types.
 pub fn router<Host, Signet>() -> Router<ctx::RpcCtx<Host, Signet>>
@@ -83,85 +81,4 @@ where
     Signet: Pnt,
 {
     ajj::Router::new().nest("eth", eth::<Host, Signet>()).nest("signet", signet::<Host, Signet>())
-}
-
-fn make_cors(cors: Option<&str>) -> CorsLayer {
-    let cors = cors
-        .unwrap_or("*")
-        .parse::<HeaderValue>()
-        .map(Into::<AllowOrigin>::into)
-        .unwrap_or_else(|_| AllowOrigin::any());
-
-    CorsLayer::new().allow_methods([Method::GET, Method::POST]).allow_origin(cors)
-}
-
-/// Serve the axum router on the specified addresses.
-async fn serve(
-    tasks: &TaskExecutor,
-    addrs: &[SocketAddr],
-    service: axum::Router,
-) -> Result<JoinHandle<()>, eyre::Error> {
-    let listener = tokio::net::TcpListener::bind(addrs).await?;
-
-    let fut = async move {
-        match axum::serve(listener, service).into_future().await {
-            Ok(_) => (),
-            Err(err) => error!(%err, "Error serving RPC via axum"),
-        }
-    };
-
-    Ok(tasks.spawn(fut))
-}
-
-/// Serve the router on the given addresses using axum.
-pub async fn serve_axum(
-    tasks: &TaskExecutor,
-    router: Router<()>,
-    addrs: &[SocketAddr],
-    cors: Option<&str>,
-) -> eyre::Result<JoinHandle<()>> {
-    let handle = tasks.handle().clone();
-    let cors = make_cors(cors);
-
-    let service = router.into_axum_with_handle("/", handle).layer(cors);
-
-    serve(tasks, addrs, service).await
-}
-
-/// Serve the router on the given address using a Websocket.
-pub async fn serve_ws(
-    tasks: &TaskExecutor,
-    router: Router<()>,
-    addrs: &[SocketAddr],
-    cors: Option<&str>,
-) -> eyre::Result<JoinHandle<()>> {
-    let handle = tasks.handle().clone();
-    let cors = make_cors(cors);
-
-    let service = router.into_axum_with_ws_and_handle("/", "/ws", handle).layer(cors);
-
-    serve(tasks, addrs, service).await
-}
-
-fn to_name(path: &std::ffi::OsStr) -> std::io::Result<ls::Name<'_>> {
-    if cfg!(windows) && !path.as_encoded_bytes().starts_with(br"\\.\pipe\") {
-        ls::ToNsName::to_ns_name::<ls::GenericNamespaced>(path)
-    } else {
-        ls::ToFsName::to_fs_name::<ls::GenericFilePath>(path)
-    }
-}
-
-/// Serve the router on the given address using IPC.
-pub async fn serve_ipc(
-    tasks: &TaskExecutor,
-    router: &Router<()>,
-    endpoint: &str,
-) -> eyre::Result<ServerShutdown> {
-    let name = std::ffi::OsStr::new(endpoint);
-    let name = to_name(name).expect("invalid name");
-    ls::ListenerOptions::new()
-        .name(name)
-        .serve_with_handle(router.clone(), tasks.handle().clone())
-        .await
-        .map_err(Into::into)
 }
